Change order of function calls
Changed what order some functions are called on the join command. This is because the whitelist check requires a guild and the join command was attempting to check the guild whitelist before it was created. The current whitelist check will be removed and replaced with discords own check function on the command in future.